+++ src/node-lib/cms/sanity-client/index.ts
@@ -21,6 +21,7 @@ import {
   supportPageSchema,
   blogListingPageSchema,
   curriculumOverviewCMSSchema,
+  newAboutWhoWeArePageSchema,
 } from "../../../common-lib/cms-types";
 import { webinarsListingPageSchema } from "../../../common-lib/cms-types/webinarsListingPage";
 import getProxiedSanityAssetUrl from "../../../common-lib/urls/getProxiedSanityAssetUrl";
@@ -90,6 +91,15 @@ const getSanityClient = () => ({
             ...whoWeArePageData,
           }
         : undefined;
+    },
+  ),
+  newAboutWhoWeArePage: getSingleton(
+    sanityGraphqlApi.newAboutWhoWeArePage,
+    newAboutWhoWeArePageSchema,
+    (result) => {
+      console.log({ result });
+      const whoWeArePageData = result?.allNewAboutCorePageWhoWeAre?.[0];
+      return whoWeArePageData;
     },
   ),
   aboutLeadershipPage: getSingleton(