+++ src/pages/about-us/get-involved.tsx
@@ -132,6 +132,7 @@ export const GetInvolved: NextPage<GetInvolvedPage> = ({ pageData }) => {
         <AboutSharedHeader
           title={pageData.header.title}
           content={pageData.header.text}
+          titleHighlight="bg-decorative3-main"
         >
           <BackgroundHeaderLoop />
         </AboutSharedHeader>